Modify quickstart to use PromptQL instead of GraphQL
Modifies the quickstart to use PromptQL instead of GraphQL.

Largely, this is a combination of the existing DDN Quickstart in conjunction with the PromptQL Quickstart. For simplicity and speed, we use the existing Docs Sample DB, but it's relatively limited in scope. We should consider a read-only dataset that's available via PG from pre-sales, possibly.