POC: Build out as a template
Builds out the PostgreSQL connector docs as a template using _partials with mdx.

There's a bug in Docusaurus + MDX that doesn't allow for interpolation of strings during build time. Instead, this happens during hydration on render. Thus, when trying to use a single partial that included
props in the headings, they rendered as {props.propName} in the right-hand ToC.