Update tutorials
Updates tutorials via a pretty big turn: we abstracted out the information and made the functions VERY generic (intentionally).

Additionally, we remove a lot of information around using custom business logic exposed as functions instead of procedures. We can re-improve this later on,
but the big focus here is the ability for custom business logic to take action on behalf of a user.