Add information about primitives
Adds information about AI primitives to the `/execute_program` docs.

These tools are all being called under the hood by PromptQL for various user-derived actions. Here, we're simply calling out that there's a (somewhat) consistent function signature in which inputs + system-generated instructions are passed and an artifact is returned. Additionally, we illustrate this with numerous examples to drive the point home while also connecting to the dots to custom business logic extending the tools available to PromptQL.

In a follow-up PR, there will be more dedicated information added around artifacts, which we'll link to from here.