Explode promptql-config section
Explodes the promptql-config section from a single page into a...section.

A major effort in this sprint is making the configuration and under-the-hood aspects of PromptQL easier to grok. Here, instead of glossing over the key-value pairs in the config, we're explicitly calling out the differences between `llm` and `ai_primitives_llm` in addition to adding richer content to the system instructions.

In a follow-up commit, we'll better shape what the best way of "speaking to" LLMs is (i.e., should we explicitly call out AI primitives when asking for task execution?).