Improve best practices by creating richer content elsewhere (#60)
## Description 📝

From poking around the source code for PromptQL and discussing this with
the ACT folks, the greatest clarity we can provide is:

1. How the configuration actually works
1. System prompts vs. Metadata descriptions

Explodes the promptql-config section from a single page into
a...section.

A major effort in this sprint is making the configuration and
under-the-hood aspects of PromptQL easier to grok. Here, instead of
glossing over the key-value pairs in the config, we're explicitly
calling out the differences between `llm` and `ai_primitives_llm` in
addition to adding richer content to the system instructions.

In a follow-up commit, we'll better shape what the best way of "speaking
to" LLMs is (i.e., should we explicitly call out AI primitives when
asking for task execution?).

Adds information for metadata descriptions and disambiguates the
differences between system instructions and metadata descriptions.

Largely, we've chosen to represent system instructions as the "behavior"
of the system and the metadata descriptions as the "how" of using each
tool.

## Quick Links 🚀

[New PromptQL Configuration
section](https://robdominguez-doc-2688-improv.promptql-docs.pages.dev/project-configuration/promptql-config/)

---------

Co-authored-by: Sean Park-Ross <94021366+seanparkross@users.noreply.github.com>