Evaluate the idea of a one click copy to llm button (#54)
## Description 📝

This PR adds the necessary logic for a one-click LLM button to our docs.
What is a one-click LLM button?

Essentially, we traverse and recurse across all our markdown files—in an
ordered fashion—and dump them into a single `.md` file which can be
copy-🍝 into an LLM by users. A follow-up PR will include the front-end
component, which will grab the raw file from our repository and add it
to the user's clipboard for them to do what they wish ❤️

There's two large buckets to this PR:
- The utility itself, written with [Bun](https://bun.sh/)
- The GH Action, which—upon merging a PR—will sneak in one last commit
updating the markdown file _only_ when the PR contains updates in the
`/docs` directory.

## Quick Links 🚀

[Raw version of the file on this
branch](https://raw.githubusercontent.com/hasura/promptql-docs/refs/heads/robdominguez/doc-2696-evaluate-the-idea-of-a-one-click-copy-to-llm-button/output/allDocs.md).

NB: I tested the above with OpenAI and Anthropic; I passed system
instructions to only use the shared docs and asked vague questions. Even
given the nature of our `mdx` and embedded components, I was impressed
by its accuracy and brevity 🎉

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>