Add frontend component for single-click llm (#55)
## Description 📝

Adds a frontend component to the navbar that reaches out to our
repository and copies the content of the most up-to-date docs. Before
merging this into `main`, we'll need to update the path (as this uses
the current branch).

## Quick Links 🚀

Available via any page, such as [the
index](https://robdominguez-docs-add-front.promptql-docs.pages.dev/index/)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>