Fix Mermaid and remove GraphiQL (#70)
## Description 📝

Fixes Mermaid and removes GraphiQL component. 

[Related issue for
Docusaurus](https://github.com/facebook/docusaurus/issues/10526)

NB: GraphiQL needed to be removed because the temporary fix of pinning
`dompurify` breaks it. It was only used in
`docs/reference/metadata-reference/relationships.mdx`. Before merge need
to confirm that that data is not generated elsewhere.

## Quick Links 🚀

<!-- Links to the relevant place(s) in the CloudFlare Pages build. Wait
for CF comment for link. -->

---------

Co-authored-by: Rob Dominguez <rob.dominguez@hasura.io>