Add an SDK section (#76)
## Description 📝

Adds an SDK section for the available SDKs (Node.js, Python, and Go)

We link to the appropriate "home" (npmjs, PyPi, GitHub), show users how
to install it, and, create a client, and how to query the available
APIs.

This is meant to be minimal as a starting place and provide a footprint
for these. Updates are expected to come from the authors going forward 🙏

## Quick Links 🚀

[New SDK
section](https://robdominguez-doc-2695-add-an.promptql-docs.pages.dev/promptql-apis/sdk/)