Clarify model providers
Clarifies model providers to use the necessary value for each service.

Additionally, feedback was raised around where to place the key...I don't think the PromptQlConfig object has made its way to LSP and the CLI yet, so holding
off on it for now. We can eventually update this page with an updated metadta object.