Mini-blitz readthrough (#123)
## Description 📝

Largely, I've elected to not make changes to the Hasura » PromptQL
language for various components and services until we have a better plan
of action.

Instead, this PR intends to take care of some small semantic nits and
more closely align the styling/state changes with what's in the
promptql.io marketing site. These changes come after conversations with
Shmo and Kai.