Docs: align documentation messaging with updated terminology from (#132)
## Description 📝

This PR has two primary goals:
- Incorporate the messaging from the OTD deck so the way we refer to
concepts is aligned with the way we're talking about PromtpQL elsewhere
- Clean up DDN references using the guidance found ☝️ 

For the latter, this typically takes the following shape:

| Old term | New term |
| --- | --- |
| supergraph | agentic\|semantic metadata layer |
| ddn [engine] | distributed query engine |
| DDN CLI | CLI |
| Hasura metadata | metadata |

> [!TIP]
> For the reviewer:
> The PR is broken into commits by each section of docs; I'd recommend
reviewing in the same fashion as it can be a lot.
>
> Additionally, I'd also recommend examining the following pages in
isolation: `docs/index.mdx` &`docs/quickstart.mdx`.

### Considerations
- The term `supergraph` is still very much necessary and accurate, but
is used more specifically now
- I avoided touching the following sections: Deployment, Private DDN,
pre-built metadata & CLI reference docs, and Policies.