Docs: Refocus quickstart on end users
Refocuses the quickstart on end users as part of the HEP: Iterating Docs to Re-Focus on Customer Personas.

At its current state, this is a `DRAFT` PR which works to succinctly highlight what happens whenever a user asks a question of PromptQL (decision making use case). In a follow-up commit,
we'll introduce the concept of automation (our other use case) and ensure we have a high-level overview between the two use cases which illustrates how users are expected to interact
with the product.