Fix: Feedback component
Fixes feedback component by modifying the endpoint.

Currently, the concatenated URL for the fetch here is `https://website-api.hasura.io/feedback/public-new-feedback?`, which returns a `404`. The service actually lives at `https://website-api.hasura.io/docs-services/docs-server/feedback/public-new-feedback`.

I've made the modification in this fetch because I'm not 100% sure where else `docsServerURL` is used and want to not cause disruptions to other components.

If this is accurate, it will need to be ported to `hasura/ddn-docs`.