Docs: Add how-to-talk section
Adds a how-to-talk to PromptQL section.

This is primarily based off work with the ACT folks and from being embedded with them in conversations to end users. This section will expand, but for now, it covers a very quick read
that we can provide as collateral to end users who are first exploring the product.