Docs: Add Architecture page
Adds a page for Architecture that's purposefully very broad.

If you take a peek at the categories in the sidebar, we enter with the quickstart for end users and then build by talking about capabilities, steering, and use cases.

This is the first page in the Core Concepts section and is meant to be as high-level of an overview as possible. We could—and should—introduce a section within this page
that highlights differences bewteen the data and control planes. But, for now, this simply serves to illustrate the various services that are running and how they work together
to produce the end-user experience.

An easy way to think of this: what's (relevant) in the root `compose.yaml` and what does each service do?