Docs: Refactor Data Modeling » Semantic Metadata
In line with our core messaging, this transforms the Data Modeling section into a Semantic Metadata section.

At first, this is a very superficial change. However, this whole category is posied to become the lynchpin between introductory end-user content and deeper knowledge for builders so they understand the relevance and importance of their metadata.